Use ListSnapshots & RestoreBackup AsyncActions on restore UI flows
Remove RestoreRequestHandler, BackupRestoreHandler and other Handler logic.
Remove BackupContext

Remove _restoreFinishedController, _multipleRestoreController, _restoreRequestController
