+++ lib/routes/initial_walkthrough/dialogs/select_backup_provider_dialog.dart
@@ -3,20 +3,25 @@ import 'package:breez/bloc/backup/backup_actions.dart';
 import 'package:breez/bloc/backup/backup_bloc.dart';
 import 'package:breez/bloc/backup/backup_model.dart';
 import 'package:breez/bloc/blocs_provider.dart';
+import 'package:breez/routes/initial_walkthrough/dialogs/restore_dialog.dart';
+import 'package:breez/routes/initial_walkthrough/loaders/loader_indicator.dart';
 import 'package:breez/routes/security_pin/remote_server_auth.dart';
+import 'package:breez/widgets/error_dialog.dart';
+import 'package:breez/widgets/flushbar.dart';
 import 'package:breez_translations/breez_translations_locales.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_easyloading/flutter_easyloading.dart';
 
 class SelectBackupProviderDialog extends StatefulWidget {
   final BackupSettings backupSettings;
   final List<BackupProvider> backupProviders;
-  final Function(BackupProvider backupProvider) onProviderSelected;
+  final Sink<bool> reloadDatabaseSink;
 
   const SelectBackupProviderDialog({
     Key key,
     this.backupSettings,
     this.backupProviders,
-    this.onProviderSelected,
+    this.reloadDatabaseSink,
   }) : super(key: key);
 
   @override
@@ -141,10 +146,105 @@ class SelectBackupProviderDialogState
     );
     backupBloc.backupActionsSink.add(updateBackupSettingsAction);
     updateBackupSettingsAction.future.then((updatedBackupSettings) {
-      Navigator.pop(context);
-      widget.onProviderSelected(selectedProvider);
+      _listSnapshots(backupBloc);
     }).catchError((err) {
       Navigator.pop(context);
     });
+  }
+
+  Future _listSnapshots(BackupBloc backupBloc) {
+    var listBackupsAction = ListSnapshots();
+
+    EasyLoading.show(
+      indicator: const LoaderIndicator(
+        message: 'Loading Backups',
+      ),
+    );
+
+    backupBloc.backupActionsSink.add(listBackupsAction);
+    return listBackupsAction.future
+        .then((snapshots) {
+          _handleRestoreRequest(snapshots);
+        })
+        .catchError((error) => _handleError(error))
+        .whenComplete(() {
+          Navigator.pop(context);
+          EasyLoading.dismiss();
+        });
+  }
+
+  void _handleError(error) {
+    EasyLoading.dismiss();
+
+    Navigator.popUntil(context, (route) {
+      return route.settings.name == "/intro";
+    });
+    if (error.runtimeType != SignInFailedException) {
+      SnackBar snackBar = SnackBar(
+        duration: const Duration(seconds: 3),
+        content: Text(error.toString()),
+      );
+      ScaffoldMessenger.of(context).showSnackBar(snackBar);
+    } else {
+      _handleSignInException(error as SignInFailedException);
+    }
+  }
+
+  Future _handleSignInException(SignInFailedException e) async {
+    if (e.provider == BackupSettings.icloudBackupProvider()) {
+      final texts = context.texts();
+      final themeData = Theme.of(context);
+
+      await promptError(
+        context,
+        texts.initial_walk_through_sign_in_icloud_title,
+        Text(
+          texts.initial_walk_through_sign_in_icloud_message,
+          style: themeData.dialogTheme.contentTextStyle,
+        ),
+      );
+    } else if (e.provider == BackupSettings.googleBackupProvider()) {
+      showFlushbar(
+        context,
+        duration: const Duration(seconds: 3),
+        message: "Failed to sign into Google Drive.",
+      );
+    }
+  }
+
+  void _handleRestoreRequest(List<SnapshotInfo> snapshots) async {
+    if (snapshots.isEmpty) {
+      _handleEmptySnapshot();
+      return;
+    }
+    _selectSnapshotToRestore(snapshots);
+  }
+
+  void _handleEmptySnapshot() {
+    Navigator.popUntil(context, (route) {
+      return route.settings.name == "/intro";
+    });
+    final texts = context.texts();
+    SnackBar snackBar = SnackBar(
+      duration: const Duration(seconds: 3),
+      content: Text(texts.initial_walk_through_error_backup_location),
+    );
+    ScaffoldMessenger.of(context).showSnackBar(snackBar);
+  }
+
+  void _selectSnapshotToRestore(
+    List<SnapshotInfo> snapshots,
+  ) async {
+    Navigator.popUntil(context, (route) {
+      return route.settings.name == "/intro";
+    });
+    showDialog<SnapshotInfo>(
+      useRootNavigator: false,
+      context: context,
+      builder: (_) => RestoreDialog(
+        snapshots,
+        widget.reloadDatabaseSink,
+      ),
+    );
   }
 }
