--- lib/routes/initial_walkthrough/handlers/models/backup_context.dart
@@ -1,12 +1,0 @@
-import 'package:breez/bloc/backup/backup_bloc.dart';
-import 'package:breez/bloc/backup/backup_model.dart';
-
-class BackupContext {
-  final BackupSettings settings;
-  final List<SnapshotInfo> snapshots;
-
-  BackupContext(
-    this.snapshots,
-    this.settings,
-  );
-}
