--- lib/routes/initial_walkthrough/handlers/models/handler_context_provider.dart
@@ -1,7 +1,0 @@
-import 'package:flutter/widgets.dart';
-
-mixin HandlerContextProvider<T extends StatefulWidget> on State<T> {
-  BuildContext getBuildContext() {
-    return context;
-  }
-}
